Refactor passkey registration logic by consolidating hooks and updating context usage; introduce usePasskeyRegistrationToast for toast management and enhance modal integration.
