+++ packages/template-retail-react-app/app/pages/registration/index.jsx
@@ -17,7 +17,7 @@ import RegisterForm from '@salesforce/retail-react-app/app/components/register'
 import useNavigation from '@salesforce/retail-react-app/app/hooks/use-navigation'
 import useEinstein from '@salesforce/retail-react-app/app/hooks/use-einstein'
 import useDataCloud from '@salesforce/retail-react-app/app/hooks/use-datacloud'
-import {usePasskeyRegistration} from '@salesforce/retail-react-app/app/hooks/use-passkey-registration'
+import {usePasskeyRegistration} from '@salesforce/retail-react-app/app/contexts/passkey-registration-provider'
 import {API_ERROR_MESSAGE} from '@salesforce/retail-react-app/app/constants'
 import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
 
@@ -32,7 +32,7 @@ const Registration = () => {
     const register = useAuthHelper(AuthHelpers.Register)
 
     const config = getConfig()
-    const {showToast} = usePasskeyRegistration()
+    const {showRegisterPasskeyToast} = usePasskeyRegistration()
 
     const submitForm = async (data) => {
         const body = {
@@ -71,7 +71,7 @@ const Registration = () => {
                     window.PublicKeyCredential.isConditionalMediationAvailable()
                 ]).then((results) => {
                     if (results.every((r) => r === true)) {
-                        showToast()
+                        showRegisterPasskeyToast()
                     }
                 })
             }
